requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
pandas==2.0.3
openpyxl==3.1.2
bcrypt==4.0.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
Werkzeug==2.3.7
Jinja2==3.1.2
numpy==1.24.3